Split xilem lib.rs into multiple files
Having all the main items in one file made that code especially hard to read and review.
Splitting across files should hopefully make that easier.
