+++ xilem/src/any_view.rs
@@ -28,7 +28,7 @@ pub type AnyWidgetView<State, Action = ()> =
 
 impl<W: Widget + FromDynWidget + ?Sized> SuperElement<Pod<W>, ViewCtx> for Pod<DynWidget> {
     fn upcast(ctx: &mut ViewCtx, child: Pod<W>) -> Self {
-        ctx.new_pod(DynWidget {
+        ctx.create_pod(DynWidget {
             inner: child.erased_widget_pod(),
         })
     }
