+++ xilem/src/view/zstack.rs
@@ -75,7 +75,7 @@ where
         for child in elements.into_inner() {
             widget = widget.with_child_pod(child.widget.erased_widget_pod(), child.alignment);
         }
-        let pod = ctx.new_pod(widget);
+        let pod = ctx.create_pod(widget);
         (pod, seq_state)
     }
