+++ xilem/src/widget_view.rs
@@ -1,0 +1,82 @@
+// Copyright 2025 the Xilem Authors
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::view::{Transformed, transformed};
+use masonry::core::{FromDynWidget, Widget};
+
+use crate::{Pod, ViewCtx};
+use masonry::kurbo::Affine;
+use xilem_core::{View, ViewSequence};
+
+use crate::any_view::AnyWidgetView;
+
+#[expect(missing_docs, reason = "TODO - Document these items")]
+pub trait WidgetView<State, Action = ()>:
+    View<State, Action, ViewCtx, Element = Pod<Self::Widget>> + Send + Sync
+{
+    type Widget: Widget + FromDynWidget + ?Sized;
+
+    /// Returns a boxed type erased [`AnyWidgetView`]
+    ///
+    /// # Examples
+    /// ```
+    /// use xilem::{view::label, WidgetView};
+    ///
+    /// # fn view<State: 'static>() -> impl WidgetView<State> + use<State> {
+    /// label("a label").boxed()
+    /// # }
+    ///
+    /// ```
+    fn boxed(self) -> Box<AnyWidgetView<State, Action>>
+    where
+        State: 'static,
+        Action: 'static,
+        Self: Sized,
+    {
+        Box::new(self)
+    }
+
+    /// This widget with a 2d transform applied.
+    ///
+    /// See [`transformed`] for similar functionality with a builder-API using this.
+    /// The return type is the same as for `transformed`, and so also has these
+    /// builder methods.
+    fn transform(self, by: Affine) -> Transformed<Self, State, Action>
+    where
+        Self: Sized,
+    {
+        transformed(self).transform(by)
+    }
+}
+
+impl<V, State, Action, W> WidgetView<State, Action> for V
+where
+    V: View<State, Action, ViewCtx, Element = Pod<W>> + Send + Sync,
+    W: Widget + FromDynWidget + ?Sized,
+{
+    type Widget = W;
+}
+
+/// An ordered sequence of widget views, it's used for `0..N` views.
+/// See [`ViewSequence`] for more technical details.
+///
+/// # Examples
+///
+/// ```
+/// use xilem::{view::prose, WidgetViewSequence};
+///
+/// fn prose_sequence<State: 'static>(
+///     texts: impl Iterator<Item = &'static str>,
+/// ) -> impl WidgetViewSequence<State> {
+///     texts.map(prose).collect::<Vec<_>>()
+/// }
+/// ```
+pub trait WidgetViewSequence<State, Action = ()>:
+    ViewSequence<State, Action, ViewCtx, Pod<crate::any_view::DynWidget>>
+{
+}
+
+impl<Seq, State, Action> WidgetViewSequence<State, Action> for Seq where
+    Seq: ViewSequence<State, Action, ViewCtx, Pod<crate::any_view::DynWidget>>
+{
+}
